+++ src/pages/Events.jsx
@@ -57,7 +57,7 @@ export default function Events() {
   }, [])
 
   const getEvents = async () => {
-    const q = query(collection(db, "events"), orderBy("eventStart", "desc"));
+    const q = query(collection(db, "events"), orderBy("eventStart", "asc"));
     const querySnapshot = await getDocs(q);
     const l = [];
     querySnapshot.forEach((doc) => {